+++ src/components/Sidebar/SidebarOption.vue
@@ -247,6 +247,7 @@ const commomProps = computed(() => {
 
   return {
     icon: props.option.icon,
+    iconSrc: props.option.iconSrc,
     iconRotate180deg: props.iconRotate180deg,
     title: props.option.label,
     label: props.option.tag,
@@ -288,8 +289,7 @@ const commomProps = computed(() => {
 
     padding-top: $unnnic-spacing-nano;
     padding-left: $unnnic-spacing-ant;
-    border-left: $unnnic-border-width-thinner solid
-      $unnnic-color-neutral-darkest;
+    border-left: $unnnic-border-width-thinner solid $unnnic-color-border-soft;
 
     box-sizing: border-box;
   }
@@ -305,8 +305,8 @@ const commomProps = computed(() => {
     left: 0;
 
     padding: $unnnic-spacing-xs;
-    border-radius: $unnnic-border-radius-sm;
-    background-color: $unnnic-color-neutral-darkest;
+    border-radius: $unnnic-radius-2;
+    background-color: $unnnic-color-border-base;
     box-shadow: $unnnic-shadow-level-near;
     width: 14.875 * $unnnic-font-size;
     box-sizing: border-box;
@@ -316,7 +316,7 @@ const commomProps = computed(() => {
     &__title {
       user-select: none;
 
-      color: $unnnic-color-neutral-clean;
+      color: $unnnic-color-fg-base;
       font-family: $unnnic-font-family-secondary;
       font-weight: $unnnic-font-weight-regular;
       font-size: $unnnic-font-size-body-md;