+++ src/components/common/RightBar/updateOrg.vue
@@ -190,7 +190,6 @@ export default {
           type: 'confirm',
           data: {
             persistent: true,
-            icon: 'alert-circle-1',
             scheme: 'feedback-red',
             title: this.$t('account.2fa.modals.disable.title'),
             description: this.$t('account.2fa.modals.org_disable.description', {
@@ -247,7 +246,6 @@ export default {
       this.openModal({
         type: 'alert',
         data: {
-          icon: 'check_circle',
           scheme: 'feedback-green',
           title: this.$t('account.2fa.modals.org_enabled.title'),
           description: this.$t('account.2fa.modals.org_enabled.description'),
@@ -259,7 +257,6 @@ export default {
       this.openModal({
         type: 'alert',
         data: {
-          icon: 'check_circle',
           scheme: 'feedback-green',
           title: this.$t('account.2fa.modals.disabled.title'),
           description: this.$t('account.2fa.modals.disabled.description'),
@@ -288,7 +285,6 @@ export default {
       this.openModal({
         type: 'alert',
         data: {
-          icon: 'check_circle',
           scheme: 'feedback-green',
           title: this.$t('orgs.delete_confirmation_title'),
           description: this.$t('orgs.delete_confirmation_text', {
@@ -302,7 +298,6 @@ export default {
       this.openModal({
         type: 'confirm',
         data: {
-          icon: 'alert-circle-1',
           scheme: 'feedback-red',
           persistent: true,
           title: this.$t('orgs.delete.title'),